backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.22
pydantic==1.10.11
python-dotenv==1.0.0
google-genai==1.29.1
httpx==0.24.1
beautifulsoup4==4.12.2
pypdf==3.17.0
gtts
pypdf
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
langdetect==1.0.9
